requirements.txt
accelerate==1.12.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
build==1.3.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
chromadb==1.4.0
click==8.3.1
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.3
cycler==0.12.1
dill==0.4.0
distro==1.9.0
docling==2.65.0
docling-core==2.57.0
docling-ibm-models==3.10.3
docling-parse==4.7.2
dotenv==0.9.9
durationpy==0.10
et_xmlfile==2.0.0
Faker==39.0.0
fastapi==0.127.0
filelock==3.20.1
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-genai==1.56.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
groq==1.0.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kubernetes==34.1.0
langchain==1.2.0
langchain-core==1.2.5
langchain-google-genai==4.1.2
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.0
latex2mathml==3.78.1
lxml==6.0.2
markdown-it-py==4.0.0
marko==2.2.1
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mmh3==5.2.0
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.18
narwhals==2.14.0
networkx==3.6.1
numpy==2.2.6
oauthlib==3.3.1
ocrmac==1.0.0
omegaconf==2.3.0
onnxruntime==1.23.2
opencv-python==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
ormsgpack==1.12.1
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
plotly==6.5.0
pluggy==1.6.0
polyfactory==3.2.0
posthog==5.4.0
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.2.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pyclipper==1.4.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pylatexenc==2.10
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-CoreML==12.1
pyobjc-framework-Quartz==12.1
pyobjc-framework-Vision==12.1
pyparsing==3.3.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.21
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
rapidocr==3.4.5
referencing==0.37.0
regex==2025.11.3
reportlab==4.4.7
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
rtree==1.4.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
semchunk==2.2.2
sentence-transformers==5.2.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
starlette==0.50.0
streamlit==1.52.2
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.3
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-java==0.23.5
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.3.0
uuid_utils==0.12.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
xlsxwriter==3.2.9
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
